+++ packages/ui/src/widgets/Leaderboard/Leaderboard.tsx
@@ -8,7 +8,6 @@ import { useAsyncCallback, useConfigContext, useMediaQuery, useWalletContext } f
 export type LeaderboardProps = Pick<TableProps, 'data'> & {
   onPlayAgain?: () => Promise<void> | void;
   onTweet?: () => Promise<void> | void;
-  serviceUrl: string;
 };
 
 const Widget = styled.div({
@@ -47,10 +46,10 @@ const TimeLeftText = styled(Typography)({
   },
 });
 
-const MysteryBlock = styled.div(({ staticBaseUrl }: { staticBaseUrl: string }) => ({
+const MysteryBlock = styled.div(({ url }: { url: string }) => ({
   height: 148,
   width: '100%',
-  background: `url(${staticBaseUrl}/Mystery_Box_1_b57b8c650d.png) no-repeat`,
+  background: `url(${url}) no-repeat`,
   backgroundSize: '90%',
   backgroundPositionY: 'center',
   marginBottom: 25,
@@ -100,7 +99,7 @@ const StyledTypography = styled(Typography)({
   letterSpacing: '0.01em',
 });
 
-export const Leaderboard = ({ data, onPlayAgain, onTweet, serviceUrl }: LeaderboardProps) => {
+export const Leaderboard = ({ data, onPlayAgain, onTweet }: LeaderboardProps) => {
   const { sessionPrice, gamePortalUrl, staticBaseUrl } = useConfigContext();
   const { address, balance = 0, isReady } = useWalletContext();
   const playerData = useMemo(() => data.find((row) => row.address === address), [data, address]);
@@ -113,14 +112,9 @@ export const Leaderboard = ({ data, onPlayAgain, onTweet, serviceUrl }: Leaderbo
 
   const handleShareClick = useCallback(async () => {
     await onTweet?.();
-    const score = data.find((row) => row.address === address)!.score;
-    const title = 'Your score';
-    const subTitle = 'Congratulations!';
-    const bgUrl = `${staticBaseUrl}/Share_twitter_6_bg_6f4253090d.png`;
-    const staticPageUrl = `${serviceUrl}/twitter-static?serviceUrl=${serviceUrl}&title=${title}&subTitle=${subTitle}&score=${score}&bgUrl=${bgUrl}`;
-    const tweetBody = `text=Do you think you can beat my Metaverse Dash Run high-score?%0a%0aMy score: ${playerData?.score}%0a%0aPlay it straight from your browser here: ${window.location.href}%0a%0a&hashtags=metaversadash,web3,gamer%0a%0a&url=${staticPageUrl}`;
+    const tweetBody = `text=Do you think you can beat my Metaverse Dash Run high-score?%0a%0aMy score: ${playerData?.score}%0a%0aPlay it straight from your browser here: ${window.location.href}%0a%0a&hashtags=metaversadash,web3,gamer`;
     window.open(`https://twitter.com/intent/tweet?${tweetBody}`, '_system', 'width=600,height=600');
-  }, [address, data, onTweet, playerData?.score, serviceUrl, staticBaseUrl]);
+  }, [onTweet, playerData?.score]);
 
   return (
     <Widget>
@@ -145,7 +139,7 @@ export const Leaderboard = ({ data, onPlayAgain, onTweet, serviceUrl }: Leaderbo
                     </StyledTypography>
                   </Stack>
                 </div>
-                <MysteryBlock staticBaseUrl={staticBaseUrl} />
+                <MysteryBlock url={staticBaseUrl.mysteryBox} />
               </Row>
               <Row columns={isMobile ? 'auto 128px' : 'auto 145px'}>
                 <GamePortalButton icon={<InsertLinkIcon />} onClick={handleOpenGamePortal}>