+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Release Notes:
 
+###vNext
+
+- Moved static base URL in the ENV config
+
 ### v0.1.2
 
 - Added tweet image